+++ desktop/src/desktop_wrapper/message_executor.rs
@@ -1,60 +1,19 @@
-use graph_craft::wasm_application_io::WasmApplicationIo;
 use graphite_editor::{application::Editor, messages::prelude::Message};
 use std::collections::VecDeque;
 
-use crate::desktop_wrapper::WgpuContext;
-use crate::desktop_wrapper::messages::{DesktopFrontendMessage, DesktopWrapperMessage};
+use super::handle_desktop_wrapper_message::handle_desktop_wrapper_message;
+use super::intercept_frontend_message::intercept_frontend_message;
+use super::intercept_message::intercept_message;
+use super::messages::{DesktopFrontendMessage, DesktopWrapperMessage};
 
-#[path = "handle_editor_message.rs"]
-mod handle_editor_message;
-
-#[path = "intercept_frontend_message.rs"]
-mod intercept_frontend_message;
-#[path = "intercept_message.rs"]
-mod intercept_message;
-
-pub struct DesktopWrapper {
-	editor: Editor,
-}
-
-impl DesktopWrapper {
-	pub fn new() -> Self {
-		Self { editor: Editor::new() }
-	}
-
-	pub fn init(&self, wgpu_context: WgpuContext) {
-		let application_io = WasmApplicationIo::new_with_context(wgpu_context);
-		futures::executor::block_on(graphite_editor::node_graph_executor::replace_application_io(application_io));
-	}
-
-	pub fn dispatch(&mut self, message: DesktopWrapperMessage) -> Vec<DesktopFrontendMessage> {
-		let mut executor = EditorMessageExecutor::new(&mut self.editor);
-		executor.queue(message);
-		executor.execute()
-	}
-
-	pub async fn run_node_graph() -> NodeGraphExecutionResult {
-		let result = graphite_editor::node_graph_executor::run_node_graph().await;
-		match result {
-			(true, texture) => NodeGraphExecutionResult::HasRun(texture.map(|t| t.texture)),
-			(false, _) => NodeGraphExecutionResult::NotRun,
-		}
-	}
-}
-
-pub enum NodeGraphExecutionResult {
-	HasRun(Option<wgpu::Texture>),
-	NotRun,
-}
-
-struct EditorMessageExecutor<'a> {
+pub(crate) struct DesktopWrapperMessageExecutor<'a> {
 	editor: &'a mut Editor,
 	queue: VecDeque<DesktopWrapperMessage>,
 	messages: Vec<Message>,
 	responses: Vec<DesktopFrontendMessage>,
 }
 
-impl<'a> EditorMessageExecutor<'a> {
+impl<'a> DesktopWrapperMessageExecutor<'a> {
 	pub(crate) fn new(editor: &'a mut Editor) -> Self {
 		Self {
 			editor,
@@ -74,7 +33,7 @@ impl<'a> EditorMessageExecutor<'a> {
 	}
 
 	pub(super) fn queue_message(&mut self, message: Message) {
-		if let Some(message) = intercept_message::intercept_message(self, message) {
+		if let Some(message) = intercept_message(self, message) {
 			self.messages.push(message);
 		}
 	}
@@ -98,7 +57,7 @@ impl<'a> EditorMessageExecutor<'a> {
 	fn process_queue(&mut self) {
 		while !self.queue.is_empty() || !self.messages.is_empty() {
 			while let Some(message) = self.queue.pop_front() {
-				handle_editor_message::handle_editor_message(self, message);
+				handle_desktop_wrapper_message(self, message);
 			}
 			let frontend_messages = self
 				.editor
@@ -106,7 +65,7 @@ impl<'a> EditorMessageExecutor<'a> {
 					messages: std::mem::take(&mut self.messages).into_boxed_slice(),
 				})
 				.into_iter()
-				.filter_map(|m| intercept_frontend_message::intercept_frontend_message(self, m))
+				.filter_map(|m| intercept_frontend_message(self, m))
 				.collect::<Vec<_>>();
 			self.respond(DesktopFrontendMessage::ToWeb(ron::to_string(&frontend_messages).unwrap().into_bytes()));
 		}
